Begin twitter media source
This is the start of the twitter media source. Right now it captures
nothing but an error screen due to some anti-scraping tech I'm assuming.
We may want to just use the oEmbed API or even the data API (to get
related information).

This also updates the entire media_source generators to add url
validation as an option and better typing with error messages.