Add full feature twitter and instagram videos. (#44)
* Update various gems for point releases

* Add image attchament to Tweet

This adds an image attchament to tweets. Well, multiple images, based
off the Instagram model.

Note: Migration required

* Add images to tweets

This adds images for tweets coming from Birdsong.

It also renames a bunch of things to allow `tweet.images` and
`instagram.images` to access the attached files. This makes it equal
across all different types of models and reduces the chance of typos.`

Migration required

* Add video to twitter and instagram

This adds video attachments to instagram and twitter. Though the
platforms don't support more than one video, for consistencey the field
is `videos` and an array. This will help with consistancy vs the images
and for other future platforms that may have more than one video.

Requries migration

* Properly access nested array when opening video files

* Rubify array element selection. [0] -> 'first'

Co-authored-by: Asa Royal <oneroyalace@gmail.com>