Remove Spring from Rails
Spring is a way to speed up repeated launches and test and such by
running all the includes and saving a forked process. This is usually
awesome! It is not good for Vips (which we use for image matching)
because it causes Vips to just hang. This also causes some issues when
debugging local gems which is a huge pain, so this fixes that as well.