Add image search (#59)
This does the following in the interest of adding reverse image search:

* Removes Spring from Rails

Spring is a way to speed up repeated launches and test and such by
running all the includes and saving a forked process. This is usually
awesome! It is not good for Vips (which we use for image matching)
because it causes Vips to just hang. This also causes some issues when
debugging local gems which is a huge pain, so this fixes that as well.

* Add dhashes to every image

This adds the `dhash-vips` gem to create a [dhash](http://www.hackerfactor.com/blog/index.php?/archives/529-Kind-of-Like-That.html) for each image before saving.

Requires migration & bundle install

* Implement very early image searching

This implements and tests instagram post image searching. The next step
is to expand this to search all images.

* Returning image search results

* Add proper loading of image search results

Image search results now properly streamed back using Turbo. Stimulus
handles the laoding, and refreshing a search page reperforms the search
and shows it properly.

* Update design for image search results

This adds better design elements for image search results, along with
a bunch of filler data.

* Finalizing image search result design

This finalizes and mixes in early mobile design for image search
results.

* Refactor Models folder

This refactors the models folder, but it's not finished yet.

* Finish refactor of models

This refactors models into folders, properly handles modules and adds
helper functions for backward compatibility.

* Finalize getting views working

This finishes up the refactoring to make sure all the views work.

You have to reset your dev databse after this.

* Adding in schmea.rb

* Add media source templating for image search

This adds image search result templates. This is useful in places where
different media sources have different names for different properties or
requires separate result templates.

This also resolves some lingering issues from refactoring models.

* Added type signatures