Add a salt to the api key hashing
This adds a salt string to any hashing done. This greatly increases
security if the database is ever breached and prevents against rainbow
table attacks.

The only thing to do is make sure `KEY_ENCRYPTION_SALT` is set in
`application.yml` or your environment variables. To generate the salt
run `rails secret` and copy it over.