#71 set up authentication (#72)
* initial commit for branch

* Add API end point to send in MediaReview

This adds a `ingest/submit_media_review` endpoint that takes in
a MediaReview object and archies it all. Just your basic POST with
a schema detailed in `public/json-schemas/media-review-schema.json`

To Test:
1. `bundle install`
1. `rails t`

* Add Test For MediaReview Saving

Adds a test to make sure media review is being properly stored

* Set up basic registration/login functionality

* Add approved attr to Users

* Allow admins to approve/ban users

* Use custom session controller

* api token auth

* Rebase on #69 mediareview submission endpoint
* Able to submit queries to mediareview endpoint with auth requisite

* Some light html

* remove letteropener dependency

* Add API Key authentication for API endpoints

This adds custom-rolled API authentication for API keys. This allows
a user to have multiple API keys per account, while also only allowing
the unhashed api key to be visible once.

This also fixes an issue where Devise was only enabled for development
mode.

* Fix issues with API key authentication

This fixes a few authentication issues in the previous push around API
keys and adds testing for it all.

* Add Crypto wrapper

* Add a salt to the api key hashing

This adds a salt string to any hashing done. This greatly increases
security if the database is ever breached and prevents against rainbow
table attacks.

The only thing to do is make sure `KEY_ENCRYPTION_SALT` is set in
`application.yml` or your environment variables. To generate the salt
run `rails secret` and copy it over.

* Filter api_key from params so we don't put them in the log

* Add seed data for admin user

This adds seed data for a default admin user. DO NOT RUN THIS IN
PRODUCTION

Also we add a bit of styling to the header

* Remove mail settings for dev

This removes some settings around the letter_opener gem, which we've
removed.

It also sets false to deliver emails in dev

* Seed db with an admin account

* Call model_from_url from correct model; add test for model_from_url method

* Remove test for relocated model_for_url method

* Rebase; change seed

Co-authored-by: Christopher Guess <cguess@gmail.com>