Switch from using Zorki locally to external host (#84)
Due to the fact that Instagram (and probably everyone else) blocks AWS,
Azure, etc. IP addresses we need to run the scraping from somewhere
else. To that end I created a server for it and am running it locally in
my apartment off a Raspberry Pi 3. So this changes the behaviour to call
the server instead of the local version of the gem.

This adds two new required environment variables:
`ZORKI_SERVER_URL` which points to the right server
`ZORKI_AUTH_KEY` which is the auth key for authenticating the requests

Everything else stays exactly the same.

To test:
1. Pull this branch
2. Run `bundle install` since the zorki gem is now removed from the
   requirements
3. Add the two environment variable however you're handling that
4. Run `rails t` and it'll pass