Upgrade the app to Rails 7.
This removes  Weback from the whole thing, and upgrades all the relevant
gems.

It also adds Redis to the development needs as Turbo requires it now
which we were planning on putting in anyways so that's fine.

ActionCable is probably broken right now, but I'll be fixing that in the
next commit.