Set up testing for callbacks
Since, in testing, we can't really do callbacks when testing against
a Hypatia instance this forces Hypatia to make the calls inline for
testing purposes.

This assumes a locally running Hypatia since it will block and break if
it's running any jobs at the time. Therefore we shouldn't ever test
against the production version of Hypatia. Hypatia will run at port 3001
from now on for testing so it doesn't interfere with this.

To test:
1. Start up a local Hypatia instance and start it up `rails s`
2. Make sure redis is up etc.
3. `rails t` is all that's necessary when Hypatia is set up with the new
instances