Fix TwitterMediaSource not creating Tweet
Due to the the new move to the Hypatia outside model we put the creating
of the actual InstgramPost etc. model on the callback function. However,
since the Twitter scraper isn't done in Hypatia this would never get
called after being spun out into a scraper job.

This introduces a new optional field for any MediaSource,
`self.runs_scraper_locally?` which if returns true (defaults to false if
not implemented in a subclass) then the result of calilng the scraper in
`ScraperJob` is used to immediately make the final object.

To test:
`rails t` should do it
`