Add ability to manage scrapes (#157)
* Use consistent whitespace in README headers

Affects #156

* Add ability to manage scrapes

This adds the ability to view, resubmit and delete scrapes that have not
been completed yet.

To test:
1. `rails db:migrate`
1. `rails test`

* Correct MacOS → macOS in README

Affects #156

* Update installation steps in the README

This commit updates the actual setup/installation steps based on my
(@reefdog) run through the process.

More changes to the README will come in a future commit, but this is the
minimum changes necessary for a new dev to get up and running without
error.

Affects #156

* Cleanup and updating of README and instructions

This commit makes a large number of minor-to-major changes to the README
based on my run through the setup instructions.

This is primarily formatting and language cleanup, with a few new notes
or suggestions.

Affects #156

* Add Redis requirement to README

Affects #156

* Fix method signature for Twitter user data export

The method sig for `TwitterUsersController.export_tweeter_data` had a
typo, and consequently the method would throw an error. This commit
repairs that.

Resolves #161

* Update index.html.erb

* Fix a bunch of type errors between Hypatia (#163)

* Fix a bunch of type errors between Hypatia

This fixes a few little bugs that caused some broken tests when scraping
primarily facebook and twitter. Why this happened when we though we had
fixed it awhile ago? Unsure.

However, all tests should fully pass after this.

To test:
`rails t`

* Fix TwitterMediaSource not creating Tweet

Due to the the new move to the Hypatia outside model we put the creating
of the actual InstgramPost etc. model on the callback function. However,
since the Twitter scraper isn't done in Hypatia this would never get
called after being spun out into a scraper job.

This introduces a new optional field for any MediaSource,
`self.runs_scraper_locally?` which if returns true (defaults to false if
not implemented in a subclass) then the result of calilng the scraper in
`ScraperJob` is used to immediately make the final object.

To test:
`rails t` should do it
`

* Add ability to manage scrapes

This adds the ability to view, resubmit and delete scrapes that have not
been completed yet.

To test:
1. `rails db:migrate`
1. `rails test`

* Update index.html.erb

* Fix routes messed up by rebase

This fixes up a few routes and paths that were screwedup by rebaseing
from master. All tests should pass now

* Fix bad route in _header

* Add colspan/label to Active Scrapes column header

This column header is intended to govern two cells below, so, let’s
colspan it.

Affects #152

Co-authored-by: Justin Reese <justin@justinreese.com>