Add asset precompilation step to setup instrux
This commit reflects a lot of learning, and tries to avoid a possible
race condition for new developers. It adds instructions to the README
that one should precompile assets before firing up the app for the first
time.

This is because the Tailwind CSS gem needs to build
`app/assets/builds/tailwind.css` before the asset pipeline runs its
manifest (or whatever).

- If you start the app with `rails s`, Tailwind CSS never fires up its
  watcher, so this file *never* gets generated, and you’re SOL.
- If you start the app with `./bin/dev`, Tailwind CSS does fire up its
  watcher and build this file, but due to the race condition with the
  asset pipeline, the file isn’t available until you stop and restart
  the app. Thus, you’re SOL on your first try (but fine after that).

We’d like to support both methods of starting the app (`rails s` and
`./bin/dev`), so the best solution for both is “just precompile after
your first installation of this project”. Boom.

Affects #187