Patch Flowbite to work with Turbo
In order to make Flowbite work with Turbo, we have to patch it to listen
for `turbo:load` rather than `DOMContentLoaded`.[^1]

This commit adds the patch-package Node package and its `postinstall`
Yarn script, along with a patchfile specific to Flowbite 1.4.3.
Developer instructions for why we’re doing this and how to update the
patchfile in the future have also been added to `DEVELOPMENT.md`.

#182

[^1]: https://github.com/themesberg/flowbite/issues/88