Disable Tailwind class-shaking in development
Tailwind likes to minimize the CSS bundle size by evaluating which
classes are actually in use by the markup and only including those.
That’s fantastic in production, but is a real stinker in development
since using any new class for the first time requires clobbering and
re-compiling the Tailwind bundle; and worse, this doesn’t happen
automatically by the watcher so you have to run `r assets:clobber` and
`r assets:precompile`. 😵

This commit configures Tailwind with a more liberal `safelist` when
running in the `development` Node environment.

#198