Add video search
This adds full reverse video search with the same interface as we
already had for images. From the user's standpoint it's the same, but
the backend is all sorts of new things.

To test:
rails db:migrate
rails t

start the server, click "image search" and then upload a video to see if
anything shows up. You may consider archiving a youtube video, and then
searching using the same video file to make sure it returns stuff.