Add S3 as backing storage
This adds the ability to use S3 as the backing storage for all of our
archived content. This is because Zeno runs on ephemeral containers and
so every restart would erase everything we have.

This adds the aws gem to support everything

This also adds three ENV variable. `USE_S3_DEV_TEST` which if set to
"true" then requires `AWS_ACCESS_KEY` and `AWS_ACCESS_SECRET` for dev
and test to test the S3 backing. Production will always require it.

To test:
- bundle install
- rails t

to test s3:
- bundle install
- Set `USE_S3_DEV_TEST` to `"true"` and set `AWS_ACCESS_KEY` and
  `AWS_ACCESS_SECRET` (get it from chris)
- rails t