Repair the retrieve_youtube_post method sig
Prior to this commit, scraping YouTube through the web interface, or via
the non-forced `create_from_url` method, didn’t complete because the
`YoutubeMediaSource#retrieve_youtube_post` method to which it tracked
back accidentally had the wrong type signature. We didn’t catch this
because all of our tests use the forced variant (`create_from_url!` maps
to `retrieve_youtube_post!`), which *did* have the correct sig.

This commit fixes the sig and verifies that scraping with
`create_from_url` works from both the Rails console and via the Zeno web
interface.

It also fixes the test that is meant to catch this, which was using the
forced version.

Resolves #211