Don’t reassign the `result` var passed to partial
Search results are passed into the Twitter list of search results as
`result`. We then determine if the result is an image or a video so we
can reach into the result and reference the archive item specifically.

However, the way we were doing this (by reassigning the local var
`result`) on one line broke the next line. In other words, when the
`result` reassignment occurred, we could no longer change on the very
next line if `has_key?(:video)`.

In the future we’ll simplify/refactor this partial anyway, but this at
least fixes the problem by assigning to a new var name.

#31