Add methods for generating normalized post attrs
This commit forces any class that uses the ArchivableItem module to
implement a `normalized_attrs_for_views` method that returns the
attributes needed to compose the generic view.

This is a compromise because we don’t actually want to coerce every
model into a normalized structure, but we do get lots of benefit from
a normalized view. The benefit of that will become apparent in
subsequent commits.

Issue #219