Format and localize archive item timestamps
Displaying timestamps such as published time and archived time in UTC
isn’t useful for users.

This commit adds a Stimulus controller for the archive items, wires up
the relevant timestamps, and uses the Day.js library to parse the
timestamp and redisplay in the user’s own time zone, and in our
preferred format.

It also converts the relevant HTML tags from DIVs to TIMEs and ensures
the `datetime` attribute still contains the ISO 8601 representation of
the timestamp.

Issue #219