Link archive items to respective author pages
This commit lets users click on the author information on each item to
link to the appropriate author page.

However, since our paths are dynamic based on post type, rather than
dynamically construct the URL-generating method name, we give the model
access to the Rails URL helpers through delegation and construct the
path as part of the normalized data structure.

Issue #219