Add jsconfig.json file
Mea culpa. In 884d546f9a0316f6a30331ccc7896af5d2becadd, I removed our
`tsconfig.json` file, thinking it was an irrelevant file that was only
relevant to TypeScript. That’s partially true — it does indicate
(incorrectly, in our case) that the project is TypeScript. But it *also*
contains configuration that our JS builder (esbuild) picks up and pays
attention to, most critically the include paths.

Since it actually *is* nice to be able to import directly from
`app/javascript` particularly, I’m re-adding this file, but as a
`jsconfig.json`. See the documentation page for `tsconfig`:
https://www.typescriptlang.org/docs/handbook/tsconfig-json.html#using-tsconfigjson-or-jsconfigjson

This is being applied as a hotfix to `master` since it’s currently
broken.