Add `lint:js` Yarn script
The difficulty of testing PR #233 exposed the need for a Yarn script
that made use of the local version of ESLint. In the npm ecosystem,
binaries are symlinked to `node_modules/.bin`, which isn’t automatically
added to the system `PATH`. IDEs and CI know to look here, and scripts
defined in `package.json` know to look here, but otherwise trying to
call packages via the CLI won’t resolve to this directory.

Linting our JS happens should happen automatically via developer’s IDE
and CI infrastructure, but when wanting to run it manually, we should
use this `yarn lint:js` script.