Fix Facebook Users duplicating
Instead of searching for current users by `facebook_id` when checking if
we should create a new user or just add a new piece of media to
a current user we were looking at `id` which, of course, never matched.

This fixes this error and also introduces `rails users:dedup_facebook`
as a rake task to fix this retroactively as well.

To test:
- Run `rails t test/models/facebook_user_test.rb`
- Run `rails users:dedup_facebook`

closes #239