Use correct controller for user registrations
We’re using our own controllers for user management that extend the
Devise built-ins, but in the routes file were still pointing straight to
the Devise one for registrations. This commit resolves that.

Functionally, this shouldn’t change anything yet. But it gets ahead of
accidentally trying to render the wrong views, which would be missing as
of a0cd22691d50b8a513cc4d2a209091a0db749a27.