Disable video preloading
The <video> element offers a `preload` attribute that tells browsers
whether we want them to optimistically preload (download before the user
has clicked play) videos or not. This is more of a hint than a rule, as
browsers will override this with their own restrictions in certain
circumstances.

The value of `preload` can be `none`, `metadata`, or `auto`, with most
browsers defaulting to `auto` under normal network conditions. That
means “yes, download the video before the user has clicked play”, which
on most of our pages -- with many videos -- means lots of unnecessary
data being delivered to the user.

`metadata` tries to download just a few % of the video, enough to grab
dimensions, codec, and other information. However, this is of limited
use to us, since we already have the `poster` instructing the browser of
the video’s dimensions, and `<source type>` instructing the MIME type.

Thus, we really want `none`, which tells the browser not to download any
of the video until the user clicks play.

Issue #258