Add auto updating to jobs console
This adds functionality to automatically update the jobs status page in
real time. It does this by adding a new ActionCable channel to monitor
scrape objects, as well as moving everything over to Stimulus styling.

This adds a few new routes to add Turbo functionality to make everything
*click* together.

To test:
- Start server using `./bin/dev`
- Start up Redis and Sidekiq
- Go to job status page
- Open new tab in app and archive a url using the manual process
- Go back to the status page and you should see a new "pill" appear
  showing new number of scrapes
- click on the pill, the table should reload

Closes #288