Fix resubmmit and delete rendering issue
This fixes an issue where resubmitting or deleting a scrape would break
the UI. This fixes that by moving it to proper partial rendering. Also
this fixes up a little thing that was bothering me with rendering of the
views itself as well.

Also flashes now render for resubmission or deleting without reloading
the page.