Add confirmation fields/methods to Applicant model
This commit migrates the `Applicant` table and model to add three new
fields for confirmation. These fields are the same one added by Devise
when using the `Devise::Confirmable` module. We aren’t actually using
that module because it’s hard to use outside of a `User` context, but we
are copying its basic design.

This commit also adds the model methods and tests for reading and
updating confirmation status, although it doesn’t add any controllers.

Issue #270