Enable Devise::Confirmable for users
Now that we have the mailer working for applicants, and as a
prerequisite for cleanly converting applicants to users, we want to
enable the `Devise::Confirmable` module and require that users have a
confirmed email address.

This commit enables and configures the module and associated routes, and
updates the messaging in the user settings screens to alerts users to
the need to confirm their new email address.

Issue #271