Add ability to set User reset password token
`Devise::Recoverable` provides a method for creating a reset-password
token on the `User` model, but it’s a protected method. We need to be
able to call it ourselves, for two reasons:

- We’re going to need to recreate some other Devise methods that call
  that method.
- We’d like to be able to test token-generation ourselves.

Both reasons require the method not to be protected, so we’re pulling it
up to where we can call it.

Issue #271