Assign default roles to new users
This commit adds the `new_user` and `insights_user` roles to users that
we create from applicants.

It doesn’t do this to all new users using an `after_create` callback on
the `User` model, as suggested by the Rolify docs, because we want more
control over when we assign these roles.

Issue #299