Use `new_user` role for "already-setup check
Prior to the existence of our roles and the `new_user` role, we were
using `user.sign_in_count` as a proxy for whether a user was already
setup, and thus whether we could send them setup instructions.

Now, we use the existence of the `new_user` role instead, as it’s more
explicitly intended for this.

Issue #299