Require being logged-out to access login page
Users don’t need to get to the login page while already logged in.

This explicitly *doesn’t* block the other actions. `create` because
blocking it causes some weird behavior (and who cares if people want to
force a new authentication, really; we just want to stop them from
accidentally stumbling in), and `destroy` because of course.

Issue #166