Restrict jobs status pages to admins
Prior to this commit, the Jobs page was hidden in the nav from non-admin
users, but if they knew the URL they could still get to it.

This commit locks that down and adds tests to verify. It also renames
the relevant test file to bring it in alignment with the actual
controller being tested, and cleans up one of the tests a bit.

Issue #166