Simplify author JSON export generation
This commit simplifies how we handle exporting author JSON data.
Previously, we had routes and method separate from the normal `show`
action, but a lot of the code was duplicated.

Now, we just route to the `show` action with the format set to JSON.
This way we reuse the same query code, have fewer route declarations,
and slimmer controllers.

This also:
- Renames the `ArchiveItem.prune_archive_items` method to
  `generate_pruned_json` for clarity
- Adds tests for ensuring JSON requests work

Issue #235