Nest ArchiveController in MediaVault namespace
This commit begins the process of nesting the Media Vault-specific
controllers within their own namespace and establishes the first pattern
for how we would do so with Insights as well.

For easier reviewing, this only converts the `ArchiveController` to the
new pattern at first. Once we’ve shaken out any issues, we’ll move the
remaining Media Vault-specific controllers in as well.

- Create a top-level `MediaVaultController`
- Nest and namespace `ArchiveController` and its tests
- Nest and namespace `ArchiveHelper`
- Nest and and update paths for the archive-related views and JavaScript
- Update routes and route names

Issue #301