Use initializer to define site-specific constants
This is a large commit that removes a bunch of inline strings for
customizing the Fact-Check Insights / MediaVault split, and replaces
them with an initializer (`SiteDefinitions`) that contains a hash for
each site. Each hash contains details about that site: shortname, title,
URL, and subdomain. We then generate accessor shorthands for getting the
site details by each of the attribute values.

This dramatically cleans up a bunch of strings littering the codebase
and reduces the opportunity for typo errors.

Issue #301