Stub out MediaVault pages
This commit adds the stubs for the Dashboard and Guide pages. It also
adds `MediaVaultController#index` explicitly to the controller, even
though we don’t call this route directly.

We also replace the `/archive` root endpoint with `/dashboard`, although
for now we leave the logic defined in the `ArchiveController`.

Issue #377