Bundle assets using asset pipeline and importmap
This large commit converts our CSS and JS bundling from Tailwind and
esbuild back to the Rails asset pipeline. It temporarily still includes
Tailwind, but only until I can finish the design implementation (#267).

With this commit, we also have no need to run Node in production. It’s
still part of the development tooling so that we can include ESLint and
its project-specific plugins and settings, but no longer needed in
production.

We already stopped using Flowbite elements in an earlier commit, so this
also gets rid of Flowbite.

- Reduce Node package dependencies down to just ESLint (and dev-only)
- Remove Node build tasks
- Replace `jsbundling-rails` and `cssbundling-rails` gems with
  `importmap-rails`
- Build our `application.css` using the asset pipeline, not Tailwind
- Update ESLint rules to disable `import/no-unresolved` since we’re not
  using any of the bundles that is intended to work with
- Update documentation and README
- Vendors the DayJS plugins that we’re using (so we aren’t hotloading
  off any CDN)
- Remove unused `jquery-rails` gem
- Remove unused `postcss` config files
- Remove unneeded Flowbite patch

Issue #290