Don’t send `callback_url` to Hypatia
Prior to this change, Zenodotus requests to Hypatia could send along a
callback URL. These overrode whatever default Zenodotus URL Hypatia was
configured with, allowing multiple Zeno apps to interact with the same
Hypatia.

In practice, we weren’t using this, and it was determined to be a
needless complication. This commit removes it.

Issue #400