Configure apps using host names, not URLs
This commit does away with the rigid assumption that our Insights and
Vault domains would be rooted at `factcheckinsights.local` in
development, and instead lets us configure them with env vars. This is
compatible with a wider array of development environments, for instance
those that need to use Pagekite. So we now have zero assumptions about
hostname format, and draw entirely from the env vars.

It also changes the name of these env vars from `_URL` suffixes to
`_HOST`, and changes the expected value accordingly to exclude protocol
and port number. Whereas before the var would have been `FOO_URL` with a
value of `http://www.example.com:80`, it is now `FOO_HOST` with a value
of `www.example.com`.

Besides being clearer and less rigid, this let us cut down on a lot of
boilerplate for coercing URLs into hostnames in out site definitions.

Many tests were updated as well.

Issue #397