Send a notification when a scrape is done
We somehow forgot to update the `scrapes` channel when a scrape is
fulfilled. This fixes that.

To test:
Unfortunately the only way to test this is to launch a Hypatia instance
and go through the whole thing of scraping and waiting for a callback.
The code that's called is already in production use, so maybe we just
say "ok" to it ;-)`

Closes #410