Create Admin controller and dashboard
This commit creates an AdminController that manages admin-area pages.

For now, it’s just a list of links to existing admin areas (e.g., jobs
status page) but it will take on more life soon.

It also updates the link in the navigation to point admins to this admin
area instead of directly to the jobs status page.

Isssue #273