Fix bug that made user menu clickable when hidden
This commit resolves an issue where the user menu items were clickable
(and obstructed clicking the items below it) even when hidden.

This is because we aren’t actually `display: none`’ing the menu. Instead
we’re reducing its opacity so that it can fade in when needed, because
there’s no pure-CSS way to fade in from `display:none`.

Problem is, even when hidden, it can still receive pointer events (hover
and click and etc.).

Now, it can only receive pointer events when the menu is open.