Use namespaced Stimulus controllers correctly
This commit resolves our misunderstanding about how to correctly nest
and namespace Stimulus controllers.

Nesting controllers within a directory like `controllers/foo/bar.js`
generates a controller named `foo--bar`. References to that controller
need to use that syntax, e.g. `data-controller="foo--bar"` or
`data-foo--bar-target="baz”`.

Without understanding how this works, we (I) were previously confused,
and were trying to just use the controller names without the namespace,
e.g. `data-controller="bar"`. When this (understandably) failed, we
thought it was because we needed to eagerLoad the controllers directly
from their directories, which of course did away with the namespacing
entirely.

This commits fixes all of that by:

- Allowing the eagerLoad to work as designed
- Correcting the references to the controller names and targets
- Organizing the importmap a little better to cluster all our own
  JavaScript pins

Resolves #431