Rename Insights user role
This commit renames the Insights user role from `insights_user` to
`fact_check_insights_user` to be in alignment with the site definition
shortname. We’re losing conciseness, gaining consistency.

What made this issue visible is my incorrect use of a Rolify magic
method to determine whether a user should be able to download the
Insights data or not. I typed `is_fact_check_insights_user?`, when the
real method was `is_insights_user?`. However, that just revealed the
inconsistency and prompted this role name change.

Resolves #433