Create a search loading spinner
The spinner defaults to `visibility:hidden` and is rendered visible only when a user uploads media. It resets to (`visibility:hidden`) when the media search results page loads. Once #456 is addressed and we no longer load a new page to show results, we'll need to manually hide this spinner on turbo load completion.

Issue #457