Update Everything For Proper MediaReview Export
This finalizes (as of now) JSON export of MediaReview, and cleans up
a bunch of tests around the same that was going sideways.

The old methods were correct for the published standard but the
published standard is quite out of date by now, so there had to be
a bunch of updates. This allows FactStream to properly throw MediaReview
at the Zeno instance and for Zeno to know what its getting.

To Test:
Honestly, just run the tests. The testing around the FactStream import
is a complicated setup and really not worth it at the moment.

Closes #482